MM-23832: Moves permissions role commands to own files.
