+++ commands/permissions.go
@@ -4,12 +4,9 @@
 package commands
 
 import (
-	"strings"
-
 	"github.com/mattermost/mattermost-server/v5/model"
 
 	"github.com/mattermost/mmctl/client"
-	"github.com/mattermost/mmctl/printer"
 
 	"github.com/spf13/cobra"
 )
@@ -17,11 +14,6 @@ import (
 var PermissionsCmd = &cobra.Command{
 	Use:   "permissions",
 	Short: "Management of permissions",
-}
-
-var RoleCmd = &cobra.Command{
-	Use:   "role",
-	Short: "Management of roles",
 }
 
 var AddPermissionsCmd = &cobra.Command{
@@ -52,48 +44,11 @@ var ShowRoleCmd = &cobra.Command{
 	RunE:       withClient(showRoleCmdF),
 }
 
-var AssignUsersCmd = &cobra.Command{
-	Use:   "assign [role_name] [username...]",
-	Short: "Assign users to role (EE Only)",
-	Long:  "Assign users to a role by username (Only works in Enterprise Edition).",
-	Example: `  # Assign users with usernames 'john.doe' and 'jane.doe' to the role named 'system_admin'.
-  permissions assign system_admin john.doe jane.doe
-  
-  # Examples using other system roles
-  permissions assign system_manager john.doe jane.doe
-  permissions assign system_user_manager john.doe jane.doe
-  permissions assign system_read_only_admin john.doe jane.doe`,
-	Args: cobra.MinimumNArgs(2),
-	RunE: withClient(assignUsersCmdF),
-}
-
-var UnassignUsersCmd = &cobra.Command{
-	Use:   "unassign [role_name] [username...]",
-	Short: "Unassign users from role (EE Only)",
-	Long:  "Unassign users from a role by username (Only works in Enterprise Edition).",
-	Example: `  # Unassign users with usernames 'john.doe' and 'jane.doe' from the role named 'system_admin'.
-  permissions unassign system_admin john.doe jane.doe
-
-  # Examples using other system roles
-  permissions unassign system_manager john.doe jane.doe
-  permissions unassign system_user_manager john.doe jane.doe
-  permissions unassign system_read_only_admin john.doe jane.doe`,
-	Args: cobra.MinimumNArgs(2),
-	RunE: withClient(unassignUsersCmdF),
-}
-
 func init() {
-	RoleCmd.AddCommand(
-		AssignUsersCmd,
-		UnassignUsersCmd,
-		ShowRoleCmd,
-	)
-
 	PermissionsCmd.AddCommand(
 		AddPermissionsCmd,
 		RemovePermissionsCmd,
 		ShowRoleCmd,
-		RoleCmd,
 	)
 
 	RootCmd.AddCommand(PermissionsCmd)
@@ -143,87 +98,6 @@ func removePermissionsCmdF(c client.Client, cmd *cobra.Command, args []string) e
 
 	if _, response = c.PatchRole(role.Id, &patchRole); response.Error != nil {
 		return response.Error
-	}
-
-	return nil
-}
-
-func showRoleCmdF(c client.Client, cmd *cobra.Command, args []string) error {
-	role, response := c.GetRoleByName(args[0])
-	if response.Error != nil {
-		return response.Error
-	}
-
-	tpl := `Name: {{.Name}}
-Display Name: {{.DisplayName}}
-Description: {{.Description}}
-Permissions: {{.Permissions}}
-{{range .Permissions}}
-  - {{.}}
-{{end}}
-{{if .BuiltIn}}
-Built in: yes
-{{else}}
-Built in: no
-{{end}}
-{{if .SchemeManaged}}
-Scheme Managed: yes
-{{else}}
-Scheme Managed: no
-{{end}}
-`
-
-	printer.PrintT(tpl, role)
-
-	return nil
-}
-
-func assignUsersCmdF(c client.Client, cmd *cobra.Command, args []string) error {
-	role, response := c.GetRoleByName(args[0])
-	if response.Error != nil {
-		return response.Error
-	}
-
-	users := getUsersFromUserArgs(c, args[1:])
-
-	for _, user := range users {
-		startingRoles := strings.Fields(user.Roles)
-		for _, roleName := range startingRoles {
-			if roleName == role.Name {
-				return nil
-			}
-		}
-
-		userRoles := append(startingRoles, role.Name)
-		_, response = c.UpdateUserRoles(user.Id, strings.Join(userRoles, " "))
-		if response.Error != nil {
-			return response.Error
-		}
-	}
-
-	return nil
-}
-
-func unassignUsersCmdF(c client.Client, cmd *cobra.Command, args []string) error {
-	users := getUsersFromUserArgs(c, args[1:])
-
-	for _, user := range users {
-		userRoles := strings.Fields(user.Roles)
-		originalCount := len(userRoles)
-
-		for i := 0; i < len(userRoles); i++ {
-			if userRoles[i] == args[0] {
-				userRoles = append(userRoles[:i], userRoles[i+1:]...)
-				i--
-			}
-		}
-
-		if originalCount > len(userRoles) {
-			_, response := c.UpdateUserRoles(user.Id, strings.Join(userRoles, " "))
-			if response.Error != nil {
-				return response.Error
-			}
-		}
 	}
 
 	return nil
